Refactor: Move interfaces and type declarations to separate files
- Create push-av-player.model.ts for Stream and NonConformingFile interfaces
- Create webrtc-stream.model.ts for WebRTCSession interface
- Create src/typings.d.ts for global video decoder type declarations

Addresses gemini code review comments.
