+++ src/typings.d.ts
@@ -1,0 +1,38 @@
+/**
+ *
+ * Copyright (c) 2026 Project CHIP Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/**
+ * Ambient type declarations for Web Video APIs that are not yet part of the
+ * standard TypeScript library typings.
+ */
+
+declare class EncodedVideoChunk {
+  constructor(chunk: any);
+}
+
+interface VideoDecoderConfig {
+  codec: string;
+  hardwareAcceleration: string;
+}
+
+declare class VideoDecoder {
+  constructor(decoder: any);
+  decode(chunk: any): void;
+  reset(): void;
+  close(): void;
+  configure(config: VideoDecoderConfig): void;
+}
